requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-speech==2.32.0
google-cloud-texttospeech==2.27.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
paho-mqtt==2.1.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pvporcupine==3.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.0
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
tkcalendar==1.6.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
